feat(LKR-19):separate models and schemas
